% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{mdKeyToControl}
\alias{mdKeyToControl}
\title{mdKeyToControl}
\usage{
mdKeyToControl(Samples, Key, command = c("add", "delete", "change"))
}
\arguments{
\item{Samples}{A data frame describing the samples, in which the column
\dQuote{Description} exists.}

\item{Key}{A column name in the \dQuote{Samples} data frame, to be used
as source for metadata.}

\item{command}{Either \dQuote{add} (default), \dQuote{delete}, or
\dQuote{"change"}.}
}
\description{
Function to create a metadata control file for \code{zenbu_upload}.
}
\details{
Note that for the initial tagging, it is recommented to pass the metadata
with the file upload in pseuto-GFF format.
}
\examples{
mdKeyToControl( data.frame(numbers=1:3, booleans=c(T, T, F), Description=c("a", "b", "c"))
              , "numbers"
              , "delete")
}
\seealso{
mdTableToControl
}
