% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{mdTableToControl}
\alias{mdTableToControl}
\title{mdTableToControl}
\usage{
mdTableToControl(Samples, Keys, command = c("add", "delete", "change"))
}
\arguments{
\item{Samples}{A data frame describing the samples, in which the column
\dQuote{Description} exists.}

\item{Keys}{Colum names in the \dQuote{Samples} data frame.}

\item{command}{Either \dQuote{add} (default), \dQuote{delete}, or
\dQuote{"change"}.}
}
\description{
Function to create a metadata control file for \code{zenbu_upload}.
}
\examples{
mdTableToControl( data.frame(numbers=1:3, booleans=c(T, T, F), Description=c("a", "b", "c"))
                , c("numbers", "booleans"))

}
\seealso{
mdKeyToControl
}
