% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safeZupload.R
\name{safeZupload}
\alias{safeZupload}
\title{safeZupload}
\usage{
safeZupload(file, name, assembly, desc, ...)
}
\description{
Safely upload a file to Zenbu.
}
\details{
Uploads a file to a Zenbu server using the zenbuUpload function.
If the file has already been uploaded with the same description,
skip it.  This is particularly useful when uploading from knitR
documents.

The parameters are passed as is to zenbuUpload.

WARNING: as long as the file is still in Zenbu's upload
cache, it will not prevent safeZupload to send it again.
Thus, safeZupload is not completely safe.
}
\seealso{
zenbuUpload
}

