% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenbuUpload.R
\name{zenbuUpload}
\alias{zenbuUpload}
\title{zenbuUpload}
\usage{
zenbuUpload(..., URL = "http://fantom.gsc.riken.jp/zenbu", verbose = FALSE,
  echo = FALSE, stdout = TRUE)
}
\arguments{
\item{...}{passed to the shell command directly}

\item{URL}{URL to the Zenbu browser
(default: http://fantom.gsc.riken.jp/zenbu).}

\item{verbose}{Print in R the shell command (default: FALSE).}

\item{echo}{Echo in shell the arguments passed (default: FALSE).}

\item{stdout}{Passed to the `system2` command.}
}
\description{
Wrapper to the zenbu_upload command-line tool of Zenbu
}
\examples{
\dontrun{
zenbuUpload("-collabs")
}
}
\seealso{
safeZupload
}

