% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{zzz}
\usage{
.onAttach(libname, packagename)
}
\description{
Before a R function provided by this package is run, it will
test if the program `zenbu_upload` is in the system PATH.
}

