% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{tableToMeta}
\alias{tableToMeta}
\title{tableToMeta}
\usage{
tableToMeta(df)
}
\arguments{
\item{df}{A data frame where each column is one source of metadata
and each row corresponds to one file to be uploaded (sorted
in the appropriate order).}
}
\description{
Outputs a metadata field in pseudo-GFF format for the
\code{zenbu_upload} command-line utility.
}
\examples{
colnames(airquality)
tableToMeta(head(airquality))

}
